Find better name for download function
