+++ packages/api/cms-api/src/dam/files/file-upload.service.ts
@@ -1,0 +1,51 @@
+import { Injectable } from "@nestjs/common";
+import FileType from "file-type";
+import fs from "fs";
+import got from "got";
+import os from "os";
+import stream from "stream";
+import { promisify } from "util";
+import { v4 as uuid } from "uuid";
+
+import { FileUploadInterface } from "./dto/file-upload.interface";
+import { FilesService } from "./files.service";
+
+const pipeline = promisify(stream.pipeline);
+
+@Injectable()
+class FileUploadService {
+    async createUploadFromUrl(url: string): Promise<FileUploadInterface> {
+        const tempDir = fs.mkdtempSync(`${os.tmpdir()}/download`);
+        const fakeName = uuid();
+        const tempFile = `${tempDir}/${fakeName}`;
+
+        if (url.substring(0, 4) === "http") {
+            await pipeline(got.stream(url), fs.createWriteStream(tempFile));
+            //TODO when downloading the file from http use mime type from reponse header
+        } else {
+            fs.copyFileSync(url, tempFile);
+        }
+
+        const fileType = await FileType.fromFile(tempFile);
+        const stats = fs.statSync(tempFile); // TODO don't use sync
+        const filename = url.substring(url.lastIndexOf("/") + 1);
+
+        return {
+            fieldname: FilesService.UPLOAD_FIELD,
+            originalname: `${filename}.${fileType?.ext}`,
+            encoding: "utf8",
+            mimetype: fileType?.mime as string,
+            size: stats.size,
+            destination: tempDir,
+            filename: fakeName,
+            path: tempFile,
+        };
+    }
+}
+
+/**
+ * @deprecated use FileUploadService#createUploadFromUrl instead
+ */
+const download = FileUploadService.prototype.createUploadFromUrl;
+
+export { download, FileUploadService };
